Don't manage package directory owner/group
The dpkg will create the splunk user and then chown this entire
directory to be owned by the splunk user.  We don't care much who owns
the file as long as it exists, but we can't write it as the splunk user
to start with since it doesn't exist.  In order to ensure we can manage
this resource properly in a single puppet run, ignore the owner and
group